refactor: category 관련 엔티티 entity/book/category 로 이동
