+++ src/main/java/shop/nuribooks/books/entity/book/category/BookCategory.java
@@ -1,4 +1,4 @@
-package shop.nuribooks.books.entity;
+package shop.nuribooks.books.entity.book.category;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -11,6 +11,7 @@ import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import shop.nuribooks.books.entity.Books;
 
 @Entity
 @AllArgsConstructor
