+++ src/main/java/shop/nuribooks/books/entity/book/category/Category.java
@@ -1,4 +1,4 @@
-package shop.nuribooks.books.entity;
+package shop.nuribooks.books.entity.book.category;
 
 import java.util.ArrayList;
 import java.util.List;
